[global] added filter for tasks to close
